% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{defaultPalettePCA3D}
\alias{defaultPalettePCA3D}
\title{Default palette}
\usage{
defaultPalettePCA3D(n = NULL, transparent = NULL, d3 = FALSE)
}
\arguments{
\item{n}{Number of colors to return}

\item{transparent}{character string which will be pasted to each color}

\item{d3}{if true, no transparency information will be added to the
colors}
}
\value{
A character vector with the color palette
}
\description{
Default set of colors for the pca3d package. This is a
colorblind-friendly palette, following the R cookbook.
}
\details{
The default palette contains 21 colors.
}

