% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{listShapes}
\alias{listShapes}
\title{Show all permissible shapes for pca3d}
\usage{
listShapes()
}
\value{
A data frame with permissible 3d shapes for plotting and their
pch counterparts is
returned invisibly.
}
\description{
Show all permissible shapes for pca3d
}
\details{
Show all permissible shapes for the functions pca3d and pca2d. The
shapes may be abbreviated using (matching is done with
\code{\link{pmatch}}.
}

