% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{makeMoviePCA}
\alias{makeMoviePCA}
\title{Create a PCA movie}
\usage{
makeMoviePCA(...)
}
\arguments{
\item{...}{Any parameters will be passed to movie3d()}
}
\value{
The value returned by movie3d()
}
\description{
Make a basic movie using movie3d()
}
\details{
A wrapper around movie3d with some default settings. Used to create a
movie on the fly.
}
\examples{
\dontrun{
data( metabo )
pca <- prcomp( metabo[,-1], scale.= TRUE )
pca3d(pca, group=metabo[,1])
makeMoviePCA()
}
}
