% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{snapshotPCA3d}
\alias{snapshotPCA3d}
\title{Save a 3D PCA snapshot}
\usage{
snapshotPCA3d(file)
}
\arguments{
\item{file}{Name of the file to save the snapshot to
data( metabo )
pca <- prcomp( metabo[,-1], scale.= TRUE )
pca3d(pca, group=metabo[,1])
snapshotPCA3d("testfile.png")}
}
\description{
Take a snapshot of the 3D PCA to a file.
}
\details{
This is just a wrapper around rgl.snapshot.
}
